+++ src/components/elements/DataGrid/Row/Draggable.jsx
@@ -1,7 +1,7 @@
 import { Draggable } from 'react-beautiful-dnd';
 import { RowHandler } from './Handler.jsx';
 
-const RowBody = ({ draggableId, index, row }) => {
+const RowDraggable = ({ draggableId, index, row }) => {
   // NOTE: We don't merge style from getRowProps with draggableProps because we
   // want full width row!!!
   // NOTE: content-open-quote to fake data and set default height for row
@@ -35,4 +35,4 @@ const RowBody = ({ draggableId, index, row }) => {
   );
 };
 
-export { RowBody };
+export { RowDraggable };
