+++ src/components/elements/DataGrid/index.js
@@ -1,5 +1,5 @@
-export * from './Cell/Body.jsx';
-export * from './Header/index.jsx';
-export * from './Row/Body.jsx';
-export * from './Cell/Header.jsx';
+export * from './Cell/Text.jsx';
+export * from './Header/Header.jsx';
+export * from './Row/Draggable.jsx';
+export * from './Header/Draggable.jsx';
 export * from './Header/Label.jsx';
